Move more early buffered lints to dyn lint diagnostics (5/N)
